Links and folder names updated
I updated the Links and therefore the directory of the folders to support both the v0.14.0 of the Vertcoin Core and the latest version of P2Pool. I also added a note that Unitus merge mining is now impossible due to the fork. I did this because I was unsure if I should delete all the content in that section.